Better logging around API failures. Able to load the music session in the native module and pass it via an event

+++ expo/modules/snow-audio-controls/android/src/main/java/com/simplepathstudios/snowgroove/audiocontrols/SnowModels.kt
@@ -95,7 +95,7 @@ data class AudioFile(
     @SerialName("artwork_url")
     var rawArtworkUrl: String? = null,
     @Field
-    var duration: Long = 0L,
+    var duration: Float = 0.0f,
     @Field
     var fingerprint: String = "",
     @Field
